refactor(#236): sortir le dispatcher de commandes du god-file, sans le toucher
`run_command` et son cortège vivaient dans `lib.rs`, 26 195 lignes. Ils
vivent maintenant dans `crates/pdo-daemon/src/run_command.rs`, 1 527. Le
déplacement est PUR : ni une ligne de logique, ni un statut, ni un corps de
réponse n'a bougé. Vérifié mécaniquement — `lib.rs` est octet pour octet
son ancienne version privée des six plages excisées, et le nouveau module
est octet pour octet ces six plages.

Le cluster est auto-contenu : `re_evaluate_after_command` n'avait que trois
appelants, tous dans `run_command` ; `ReEvalSummary`, `ReEvalTerminal`,
`collect_yaml_var_refs` et `RunCommandRequest` n'en avaient aucun ailleurs.

Coût en visibilité : **deux** annotations, pas une comme prévu. `run_command`
devient `pub(crate)` parce que le routeur vit dans la racine, donc dans un
ANCÊTRE ; et `RunCommandRequest` doit suivre, parce que l'extracteur axum le
hisse dans la signature publique du handler. Les 23 items racine et 12 items
de modules frères dont le cluster dépend n'ont, eux, rien exigé : Rust rend
un item racine visible dans tous ses descendants.

Les deux tests unitaires d'`extract_variable_refs_from_outgoing_edges`
suivent le helper — `crate::tests` est un frère, pas un descendant, donc les
laisser aurait coûté une troisième annotation.

Effet de bord assumé dans `lib.rs` : `spawn_node`, `SpawnDeps` et
`SpawnOutcome` n'y ont plus d'appelant de production (ils sont partis avec
`restart_node`), leur import passe donc sous `#[cfg(test)]` — même forme que
`create_sub_worktree` juste au-dessus.
